lists stay bolded when selected
