--- public/javascripts/searchbar.js
@@ -1,1 +1,0 @@
-
