+++ views/utils.pug
@@ -17,3 +17,5 @@ mixin field(fieldText, fieldName, fieldValue, fieldType = 'text', placeholder = 
 mixin plainField(fieldText, fieldName, fieldValue, fieldType = 'text', placeholder = null, autocomplete = 'off')
   label(for= fieldName) #{fieldText}
     input(class='inp-field' size=50 type=fieldType id=fieldName name=fieldName value=fieldValue placeholder=placeholder autocomplete=autocomplete)
+
+
